To generic ET, added left-multiply by general scalar type, not just float Delta S=1 anomalous dimension, fixed limit issue for pure QCD running in 3f Fixed ValOrd constructor taking int not double To K->pipi analysis, added computation of LL factor using Colangelo curve at physical pion mass Also in K->pipi analysis fixed definition of renormalized matrix elements to not include the G_F V_us V_ud/sqrt(2) coefficient

+++ include/physics/valord.h
@@ -29,7 +29,7 @@ struct ValOrd{
   ValOrd(){
     zero();
   }
-  ValOrd(const int powas, const int powa, const int val): ValOrd(){
+  ValOrd(const int powas, const int powa, const double val): ValOrd(){
     this->operator()(powas,powa) = val;
   }
